add pagination and change add,update and delete in Artist

+++ src/main/java/recordstore/service/ArtistServiceImpl.java
@@ -1,10 +1,15 @@
 package recordstore.service;
 
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 import recordstore.entity.Artist;
 import recordstore.projections.ArtistProjection;
 import recordstore.repository.ArtistRepository;
+import recordstore.utils.FileUploadUtil;
 
+import java.io.IOException;
 import java.util.List;
 
 @Service
@@ -17,13 +22,26 @@ public class ArtistServiceImpl implements ArtistService {
     }
 
     @Override
-    public void saveArtist(Artist artist) {
+    public void saveArtist(Artist artist) throws IOException {
+        String filename = StringUtils.cleanPath(artist.getData().getOriginalFilename());
+        String removeImg = artist.getImg();
+        if (!artist.getData().isEmpty()) {
+            artist.setImg(filename);
+        }
         repository.save(artist);
+        if(!artist.getData().isEmpty()) {
+            FileUploadUtil.saveFile(filename, artist.getData());
+            FileUploadUtil.deleteFile(removeImg);
+        }
     }
 
     @Override
-    public void deleteArtist(long id) {
-        repository.deleteById(id);
+    public void deleteArtist(long id) throws IOException {
+        Artist artist = repository.getOne(id);
+        if (artist.getReleases().size() == 0) {
+            repository.deleteById(id);
+            FileUploadUtil.deleteFile(artist.getImg());
+        }
     }
 
     @Override
@@ -32,8 +50,13 @@ public class ArtistServiceImpl implements ArtistService {
     }
 
     @Override
-    public List<ArtistProjection> getAllArtists() {
+    public List<ArtistProjection> getAllArtistsNames() {
         return repository.findAllBy();
+    }
+
+    @Override
+    public Page<Artist> getAllArtists(Pageable pageable) {
+        return repository.findAll(pageable);
     }
 
 //    @Override
